chore: moved ui render function to features folder
