+++ events/interaction/features/projectHelper/links/link-select-menu-handler.js
@@ -1,4 +1,4 @@
-import { renderLinkManager } from "../../../../../utils/renderLinkManager.js";
+import { renderLinkManager } from "../../../../../features/projects/links/renderLinkManager.js";
 
 export async function projectLinkSelectHandler(interaction) {
   try {
